+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/store/ArrayStore.java
@@ -2,7 +2,6 @@ package uk.ac.bbsrc.tgac.miso.core.store;
 
 import java.io.IOException;
 import java.util.List;
-import java.util.Map;
 
 import uk.ac.bbsrc.tgac.miso.core.data.Array;
 import uk.ac.bbsrc.tgac.miso.core.data.ArrayModel;
@@ -18,12 +17,6 @@ public interface ArrayStore extends Store<Array>, PaginatedDataSource<Array> {
   public ArrayModel getArrayModel(long id) throws IOException;
 
   public List<ArrayModel> listArrayModels() throws IOException;
-
-  /**
-   * @return a map containing all column names and max lengths from the Array table
-   * @throws IOException
-   */
-  public Map<String, Integer> getArrayColumnSizes() throws IOException;
 
   public List<Sample> getArrayableSamplesBySearch(String search) throws IOException;